pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0"]

[project]
authors = [
  {name = "Felix Stark", email = "felix.stark@kfw.de"},
  {name = "Ingmar Mueller", email = "ingmar.mueller@kfw.de"},
  {name = "Mario Macai", email = "mario.macai@accenture.com"},
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = ["goechargerv2"]
description = "This is an integration for the Home Assistant for the go-e Charger Cloud wallbox (API v2)."
keywords = ["goecharger", "kfw", "homeassistant", "cloud"]
name = "homeassistant_goechargerv2"
readme = "README.md"
requires-python = ">=3.7"
version = "0.1.0"

[project.optional-dependencies]
dev = ["black", "pylint", "pytest-homeassistant-custom-component", "Flask", "pre-commit"]

[project.urls]
"Bug Tracker" = "https://github.com/openkfw/homeassistant_goechargerv2/issues"
"Homepage" = "https://github.com/openkfw/homeassistant_goechargerv2"
